% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aging-data.R
\docType{data}
\name{aging}
\alias{aging}
\title{Residual DNA-methylation-based (DNAm) gestational age}
\format{
A data frame vector with 10 rows and 1 column:
\describe{
\item{output}{A numeric value of residual DNAm gestational age (weeks).}
}
}
\source{
Derived from the 2024 Placental Clock DREAM Challenge.
}
\usage{
aging
}
\description{
A data frame of DNAm gestational age for both case and control groups.
}
\keyword{dataset}
