% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_v_indices-data.R
\docType{data}
\name{beta_v_indices}
\alias{beta_v_indices}
\title{Indices for random beta values}
\format{
A list of 2 elements where each has a length of nfit of 10000:
\describe{
\item{beta1.v}{An integer indicating the selected indices.}
\item{beta2.v}{An integer indicating the selected indices.}
}
}
\source{
Derived from \code{ChAMP} R package.
}
\usage{
beta_v_indices
}
\description{
A list of indices for \code{rand.idx} in \code{\link{bmiq_norm_450k}} function.
The input probes must be already filtered and ordered the same way to the
that when we developed our placental epigenetic clock. Run
\code{data(probe_info_450k)} and find the required probes in
\code{prob_info_450k$probeID}.
}
\keyword{dataset}
