% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_values_case-data.R
\docType{data}
\name{beta_values_case}
\alias{beta_values_case}
\title{Beta values at individual CpG sites for case group}
\format{
A data frame with 452,453 rows and 5 columns:
\describe{
\item{GSM1931565}{Beta values for sample GSM1931565.}
\item{GSM5114811}{Beta values for sample GSM5114811.}
\item{GSM2589558}{Beta values for sample GSM2589558.}
\item{GSM1842848}{Beta values for sample GSM1842848.}
\item{GSM1843045}{Beta values for sample GSM1843045.}
}
}
\source{
Derived from the 2024 Placental Clock DREAM Challenge.
}
\usage{
beta_values_case
}
\description{
A data frame of beta values for 5 samples and 452,453 probes.
}
\keyword{dataset}
