% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_values_control-data.R
\docType{data}
\name{beta_values_control}
\alias{beta_values_control}
\title{Beta values at individual CpG sites for control group}
\format{
A data frame with 452,453 rows and 5 columns:
\describe{
\item{GSM7115144}{Beta values for sample GSM7115144.}
\item{GSM1702248}{Beta values for sample GSM1702248.}
\item{GSM3179749}{Beta values for sample GSM3179749.}
\item{GSM4281756}{Beta values for sample GSM4281756.}
\item{GSM5210472}{Beta values for sample GSM5210472.}
}
}
\source{
Derived from the 2024 Placental Clock DREAM Challenge.
}
\usage{
beta_values_control
}
\description{
A data frame of beta values for 5 samples and 452,453 probes.
}
\keyword{dataset}
