% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_beta_values_case-data.R
\name{download_beta_values_case}
\alias{download_beta_values_case}
\title{Access Beta values at individual CpG sites for case group}
\source{
Derived from the 2024 Placental Clock DREAM Challenge.
}
\usage{
download_beta_values_case()
}
\value{
A data frame with 452,453 rows and 5 columns:
\describe{
\item{GSM1931565}{Beta values for sample GSM1931565.}
\item{GSM5114811}{Beta values for sample GSM5114811.}
\item{GSM2589558}{Beta values for sample GSM2589558.}
\item{GSM1842848}{Beta values for sample GSM1842848.}
\item{GSM1843045}{Beta values for sample GSM1843045.}
}
}
\description{
Downloads and loads beta values for the case group. Data contains beta
values for 5 samples and 452,453 probes.
}
\examples{

beta_values_case <- download_beta_values_case()
head(beta_values_case)
}
\keyword{dataset}
