% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_beta_values_control-data.R
\name{download_beta_values_control}
\alias{download_beta_values_control}
\title{Access Beta values at individual CpG sites for control group}
\source{
Derived from the 2024 Placental Clock DREAM Challenge.
}
\usage{
download_beta_values_control()
}
\value{
A data frame with 452,453 rows and 5 columns:
\describe{
\item{GSM7115144}{Beta values for sample GSM7115144.}
\item{GSM1702248}{Beta values for sample GSM1702248.}
\item{GSM3179749}{Beta values for sample GSM3179749.}
\item{GSM4281756}{Beta values for sample GSM4281756.}
\item{GSM5210472}{Beta values for sample GSM5210472.}
}
}
\description{
Downloads and loads beta values for the control group. Data contains beta
values for 5 samples and 452,453 probes.
}
\examples{

beta_values_control <- download_beta_values_control()
head(beta_values_control)
}
\keyword{dataset}
