% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga-data.R
\docType{data}
\name{ga}
\alias{ga}
\title{Gestational age for both case and control groups}
\format{
A data frame vector with 10 rows and 1 column:
\describe{
\item{GA}{A numeric value of gestational age (weeks' gestation).}
}
}
\source{
Derived from the 2024 Placental Clock DREAM Challenge.
}
\usage{
ga
}
\description{
A data frame of gestational age for 10 samples.
}
\keyword{dataset}
