% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype-data.R
\docType{data}
\name{phenotype}
\alias{phenotype}
\title{Phenotype for both case and control groups}
\format{
A data frame vector with 10 rows and 1 column:
\describe{
\item{phenotype}{A character value of phenotype (case/control).}
}
}
\source{
Derived from the 2024 Placental Clock DREAM Challenge.
}
\usage{
phenotype
}
\description{
A data frame of phenotype for 10 samples.
}
\keyword{dataset}
