% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probe_info_450k-data.R
\docType{data}
\name{probe_info_450k}
\alias{probe_info_450k}
\title{Probe info for 450K}
\format{
A list of 2 elements where each has a length of 346407:
\describe{
\item{Design}{An integer indicating design type 1 or 2.}
\item{probeID}{A character for each probe identifier.}
}
}
\source{
Derived from \code{ChAMP} R package.
}
\usage{
probe_info_450k
}
\description{
A list of 450K probe information for \code{\link{bmiq_norm}} function. The
probes are already filtered and ordered the same way to the input when we
developed our placental epigenetic clock.
}
\keyword{dataset}
